+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/CallListRepository.kt
@@ -4,6 +4,8 @@ import pt.arbitros.arbnet.domain.CallList
 import java.time.LocalDate
 
 interface CallListRepository {
+
+
     fun createCallList(
         deadline: LocalDate,
         userId: Int,
@@ -27,19 +29,10 @@ interface CallListRepository {
         callType: String,
     ): CallList
 
-//    fun findCallListById(id: Int): CallList?
-//
+
 //    fun getCallListsByCouncil(councilId: Int): List<CallList>
 //
 //    fun deleteCallList(id: Int): Boolean
-//
-//    fun updateCallList(
-//        id: Int,
-//        deadline: LocalDate,
-//        callType: String,
-//        userId: Int,
-//        competitionId: Int,
-//    ): Boolean
-//
-//    fun getCallListReferees(callListId: Int): List<Referee>
+
+
 }
