+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/Transaction.kt
@@ -1,8 +1,11 @@
 package pt.arbitros.arbnet.repository
 
-import pt.arbitros.arbnet.repository.CategoryDirRepository
-import pt.arbitros.arbnet.repository.CategoryRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.CategoryRepository
 import pt.arbitros.arbnet.repository.auxiliary.SessionRefereesRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.EquipmentRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.FunctionRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.PositionRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.RoleRepository
 
 interface Transaction {
     val usersRolesRepository: UsersRolesRepository
