+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/jdbi/CategoryRepositoryJdbi.kt
@@ -2,9 +2,8 @@ package pt.arbitros.arbnet.repository.jdbi
 
 import org.jdbi.v3.core.Handle
 import org.jdbi.v3.core.kotlin.mapTo
-import org.springframework.context.annotation.ComponentScan
 import pt.arbitros.arbnet.domain.universal.Category
-import pt.arbitros.arbnet.repository.CategoryRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.CategoryRepository
 
 class CategoryRepositoryJdbi(
     private val handle: Handle
