+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/jdbi/EquipmentRepositoryJdbi.kt
@@ -2,7 +2,7 @@ package pt.arbitros.arbnet.repository.jdbi
 
 import org.jdbi.v3.core.Handle
 import pt.arbitros.arbnet.domain.universal.Equipment
-import pt.arbitros.arbnet.repository.EquipmentRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.EquipmentRepository
 
 class EquipmentRepositoryJdbi(
     private val handle : Handle,
