+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/jdbi/FunctionRepositoryJdbi.kt
@@ -2,7 +2,7 @@ package pt.arbitros.arbnet.repository.jdbi
 
 import org.jdbi.v3.core.Handle
 import org.jdbi.v3.core.kotlin.mapTo
-import pt.arbitros.arbnet.repository.FunctionRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.FunctionRepository
 
 class FunctionRepositoryJdbi(
     private val handle: Handle,
