+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/jdbi/PositionRepositoryJdbi.kt
@@ -3,7 +3,7 @@ package pt.arbitros.arbnet.repository.jdbi
 import org.jdbi.v3.core.Handle
 import org.jdbi.v3.core.kotlin.mapTo
 import pt.arbitros.arbnet.domain.universal.Position
-import pt.arbitros.arbnet.repository.PositionRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.PositionRepository
 
 class PositionRepositoryJdbi (
     private val handle: Handle
