+++ code/jvm/src/test/kotlin/pt/arbitros/arbnet/repository/mem/CategoryRepositoryMem.kt
@@ -1,7 +1,7 @@
 package pt.arbitros.arbnet.repository.mem
 
 import pt.arbitros.arbnet.domain.universal.Category
-import pt.arbitros.arbnet.repository.CategoryRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.CategoryRepository
 
 class CategoryRepositoryMem: CategoryRepository {
     override fun createCategory(category: Category): Int {
