+++ code/jvm/src/test/kotlin/pt/arbitros/arbnet/repository/mem/RoleRepositoryMem.kt
@@ -1,7 +1,7 @@
 package pt.arbitros.arbnet.repository.mem
 
 import pt.arbitros.arbnet.domain.universal.Role
-import pt.arbitros.arbnet.repository.RoleRepository
+import pt.arbitros.arbnet.repository.adaptable_repos.RoleRepository
 
 class RoleRepositoryMem : RoleRepository {
     private val roles = mutableMapOf<Int, String>()
